Start using a monadic error handling style
The parsers and printers will generate fatal errors when they
fail; i.e. if a client sends a bogus packet then we will generate
a loggable message and abort the connection.

This patch creates a single error type containing the loggable
error message and switches to use a monadic style in the `Request`
parser (basically to avoid nesting the code too much)

Signed-off-by: David Scott <dave@recoil.org>