`Data.t` is now `SERIALISABLE`
The only difference between the old signature and the `SERIALISABLE`
one was the read function used to return a pair including the remaining
unconsumed buffer. There is no need for this: the caller can compute
the same by calling `sizeof`.

This patch also switches to a monadic error handling style.

Signed-off-by: David Scott <dave@recoil.org>