Add a notion of `recommended` tags and fids
The client gets to choose which tags are in-use, where each tag
is a concurrent RPC and each fid corresponds to some file state on
the server. Although it's legal to use lots and lots of these
values, it seems polite to restrict ourselves to a subset of the
available tag- and fid- space.

This patch adds a notion of a `recommended` set of tags and fid
values. Our client will be nice and only use tags and fids from this
space. For now we will use only 100 tags and 100 fids.

Signed-off-by: David Scott <dave.scott@unikernel.com>