Improve type safety and error reporting of Handler
Before, any method could return any response type. Now, a filesystem
function must return the appropriate response type for the request, or
an error to be returned to the client.

Errors are converted to the appropriate format based on the protocol
version (errno is suppressed if the format does not include it, and
a default is added if errno is needed and not present).