Support request flushing
This patch adds support for simple request flushing, as required by the
Linux 9P client.

When a request arrives, a `bool ref` cell is created and associated with
the request tag.

When a `flush` arrives, if the `bool ref` can be found it is toggled to
`true`.

When the original response is to be written, if the `bool ref` is set to
`true` then the packet is dropped.

Note the scope of the `write_lock` has been widened: previously it only
protected the channel. It is now held while the cancel `bool ref` is
interrogated and while the packet is marshalled too.

Signed-off-by: David Scott <dave.scott@unikernel.com>