Avoid marshalling every packet to a string when not in debug mode
The `Log.debug` function will create a string to be logged via

```
Fmt.kstrf (fun s -> if !print_debug then print_endline s) fmt
```

before deciding whether to print them or not. In debug mode we print every
packet received and transmitted. This patch ensures that, when not in
debug mode, we don't generate vast quantities of strings which we never
print.

Signed-off-by: David Scott <dave.scott@unikernel.com>