Server: call `receive_cb` on initial `Attach`
Before this patch we were in an inconsistent situation where the first
`attach` was handled magically by the server functor, while all subsequent
`attach` requests had to be handled by the user module. This led to the
initial `root_fid` being special-cased.

This patch requires the user module to handle all `attach` requests, making
it easier for the user module to maintain a table of active fids (i.e.
no need for special-casing the `root_fid`)

Signed-off-by: David Scott <dave.scott@unikernel.com>