Flow_lwt_unix: make `close` idempotent for safety
It's a complete disaster to call `close` on an fd twice, since a
parallel thread may have just reallocated it. Since this bug is really
hard to track down, attempt to prevent it by making the higher-level
`Flow_lwt_unix.close` function idempotent by adding a `mutable bool`.

As a side-effect, `read` `write` and `writev` should fail with
`Eof` if called on a closed flow.

Signed-off-by: David Scott <dave.scott@docker.com>