Client: add `on_disconnect`
Before this patch it was impossible for users to find out if the
connection had been closed by the remote. The dispatcher thread would
discover the failure, but the client had no way to find out.

Users can now bind to `on_disconnect` and run code when the dispatcher
has disconnected. In the case of the Unix flow-based client, they
can also call `disconnect` which will have the side-effect of disconnecting
the flow.

Signed-off-by: David Scott <dave.scott@docker.com>