Client: use a per-connection condition variable for free_tags
Before this patch the "wait for a free" tag mechanism was completely
broken as each call to `rpc` would create its own fresh condition
variable (!) and then signal itself with it.

This patch uses a per-connection condition variable `free_tags_c`
to go with the per-connection set of free tags `free_tags`.

Signed-off-by: David Scott <dave.scott@docker.com>