Client: on disconnect, wake up threads blocked waiting for a fid
Before this patch requests waiting for a *tag* would be woken because
the threads currently blocked who have the tags allocated would deallocate
the tags on the way out. However requests waiting for a *fid* would
never be woken, because deallocation (aka "clunk") is a separate Client
API call which we don't expect clients to have to call once they receive
a disconnection error.

This patch explicitly wakes up all threads blocked waiting for a free
fid, causing them to see that the dispatcher thread has shutdown and
return an error.

Note the signature of `with_fid` has expanded to include an `Error.t`
type constructor to reflect this new error condition.

Signed-off-by: David Scott <dave.scott@docker.com>