Client: wake up threads on all paths out of the dispatcher
Before this patch the blocked threads were only woken up if the
`disconnect` function is called, which won't happen if the dispatcher
shuts down because of an I/O failure.

This thread ensures that under all circumstances, the dispatcher
exits and wakes up all blocked threads.

This patch clarifies the handling of the `wakeners` map: wakeners
remain in the map until they are woken, at which point they are removed
to prevent any possibility of illegal double-waking.

Signed-off-by: David Scott <dave.scott@docker.com>