Client: deallocate_fid should clunk before marking fid as free
Before this patch the fid was returned to the free pool before the
`clunk` request was created. It may have been possible for another
thread to allocate the fid and start using it only to have it
mysteriously `clunked` under their feet.

This patch keeps the fid locally allocated until the `clunk` has been
processed.

Signed-off-by: David Scott <dave@recoil.org>