Remove a copy when reading packets
We always read the 4 byte packet length first, and then read the
payload. If the underlying FLOW manages to return a contiguous
buffer, avoid copying it only to put the 4 bytes back on the front.

Note that no-one really wanted the 4 bytes back on the front since
the length is obvious by looking at the buffer length.

Signed-off-by: David Scott <dave@recoil.org>